Centralize common test code
