+++ test/Common/TestUtils/EventCollector.cs
@@ -1,11 +1,13 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace EliteFiles.Tests.Internal
+namespace TestUtils
 {
+    [SuppressMessage("Performance", "CA1812:Avoid uninstantiated internal classes", Justification = "Shared test code.")]
     internal sealed class EventCollector<T>
     {
         private readonly Action<EventHandler<T>> _attach;
@@ -17,6 +19,66 @@ namespace EliteFiles.Tests.Internal
             _attach = attach;
             _detach = detach;
             _name = name;
+        }
+
+        public T? Wait(Action trigger, int timeout = Timeout.Infinite)
+        {
+            T? res = default;
+
+            using (var ss = new SemaphoreSlim(0, 1))
+            {
+                void Handler(object? sender, T e)
+                {
+                    res = e;
+                    ss.Release();
+                }
+
+                _attach(Handler);
+                trigger();
+                bool ok = ss.Wait(timeout);
+                _detach(Handler);
+
+                if (!ok)
+                {
+                    throw new TimeoutException();
+                }
+            }
+
+            return res;
+        }
+
+        public IList<T> Wait(int count, Action trigger, int timeout = Timeout.Infinite)
+        {
+            var res = new List<T>();
+
+            using (var ce = new CountdownEvent(count))
+            {
+                void Handler(object? sender, T e)
+                {
+                    res.Add(e);
+
+                    if (ce.IsSet)
+                    {
+                        string list = string.Join(',', res.Select(x => $"{x}"));
+                        throw new InvalidOperationException($"More than {count} events received in collector '{_name}': {list}.");
+                    }
+
+                    ce.Signal();
+                }
+
+                _attach(Handler);
+                trigger();
+                bool ok = ce.Wait(timeout);
+                _detach(Handler);
+
+                if (!ok)
+                {
+                    string list = string.Join(',', res.Select(x => $"{x}"));
+                    throw new TimeoutException($"Timeout in collector '{_name}' after receiving the following events: {list}.");
+                }
+            }
+
+            return res;
         }
 
         public async Task<T?> WaitAsync(Action trigger, int timeout = Timeout.Infinite)
