+++ test/EliteChroma.Core.Tests/ModifierKeysWatcherTests.cs
@@ -6,6 +6,7 @@ using System.Xml.Linq;
 using EliteChroma.Core.Tests.Internal;
 using EliteChroma.Elite.Internal;
 using EliteFiles.Bindings;
+using TestUtils;
 using Xunit;
 using static EliteChroma.Core.Internal.NativeMethods;
