+++ test/EliteFiles.Tests/EliteFiles.Tests.csproj
@@ -35,6 +35,9 @@
     <None Include="..\Common\TestFiles\**" CopyToOutputDirectory="PreserveNewest">
       <Link>TestFiles\%(RecursiveDir)%(Filename)%(Extension)</Link>
     </None>
+    <Compile Include="..\Common\TestUtils\**" CopyToOutputDirectory="Never">
+      <Link>TestUtils\%(RecursiveDir)%(Filename)%(Extension)</Link>
+    </Compile>
   </ItemGroup>
 
   <ItemGroup>
